+++ src/main/java/com/MovieTicket/MovieBooking/service/MovieService.java
@@ -1,6 +1,8 @@
 package com.MovieTicket.MovieBooking.service;
-import com.MovieTicket.MovieBooking.model.*;
-import com.MovieTicket.MovieBooking.repository.*;
+
+import com.MovieTicket.MovieBooking.exception.GlobalExceptionHandler;
+import com.MovieTicket.MovieBooking.model.Movie;
+import com.MovieTicket.MovieBooking.repository.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
@@ -14,7 +16,7 @@ public class MovieService {
     private MovieRepository movieRepository;
 
     public List<Movie> getAllMovies() {
-    	return movieRepository.findAll(Sort.by(Sort.Direction.DESC, "id"));
+        return movieRepository.findAll(Sort.by(Sort.Direction.DESC, "id"));
     }
 
     public void saveMovie(Movie movie) {
@@ -22,21 +24,31 @@ public class MovieService {
     }
 
     public Movie getMovieById(Long id) {
-        return movieRepository.findById(id).orElse(null);
+        return movieRepository.findById(id)
+                .orElseThrow(() ->
+                        new GlobalExceptionHandler.MovieNotFoundException(
+                                "Movie not found with id: " + id
+                        )
+                );
     }
 
     public void deleteMovie(Long id) {
+        if (!movieRepository.existsById(id)) {
+            throw new GlobalExceptionHandler.MovieNotFoundException(
+                    "Movie not found with id: " + id
+            );
+        }
         movieRepository.deleteById(id);
     }
-    
+
     public long getTotalMoviesCount() {
         return movieRepository.count();
     }
-    
+
     public List<Movie> getMoviesByGenre(String genre) {
         return movieRepository.findByGenre(genre);
     }
-    
+
     public List<Movie> getNowShowing() {
         return movieRepository.findNowShowing();
     }
